# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Together API",
    "description": "Private, non-federated Discord alternative for small gaming communities",
    "contact": {
      "name": "Together Contributors"
    },
    "license": {
      "name": "PolyForm Noncommercial 1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/link-preview": {
      "get": {
        "tags": [
          "LinkPreview"
        ],
        "summary": "GET /link-preview?url=<encoded-url>",
        "description": "Returns Open Graph metadata for the given URL, with results cached for 24 hours.\nRequires authentication. Rejects private/loopback IPs (SSRF protection).",
        "operationId": "get_link_preview",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link preview metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkPreviewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LinkPreviewDto": {
        "type": "object",
        "description": "Open Graph metadata returned by `GET /link-preview`.\n\nAll fields except `url` are optional — a page may have no OG tags.",
        "required": [
          "url"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "site_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}